fix(exec-approvals): handle same-process sync/async lock contention
The exec-approvals store uses two lock mechanisms sharing the same
`.lock` file: an async sidecar lock (via @openclaw/fs-safe) and a sync
custom lock. When the async lock holds the file, the sync lock detects
ownerPid === process.pid and immediately throws file_lock_timeout
instead of retrying (to avoid self-deadlock). This triggers
createFailClosedExecApprovalsFallback() → security: deny, wedging all
subsequent exec calls.

Fix: In withExecApprovalsReadLockSync, catch file_lock_timeout errors
caused by same-process contention and fall back to running the read
function without the lock. This is safe because:

- Async writes use atomic temp-file + rename, so concurrent reads
  always see a consistent snapshot
- Only applies to READ locks; write serialization is preserved
- Falls back only when the target file already exists, preserving
  fail-closed behavior during initial policy creation

Closes #106777
